+++ include/backend/value.h
@@ -1,13 +1,15 @@
 #ifndef VALUE_H
 #define VALUE_H
 
+#include <stdint.h>
+
 /**@desc clox value*/
 typedef double Value;
 
 /**@desc dynamic array used for storing clox values*/
 typedef struct {
   Value *values;
-  long capacity, count;
+  int32_t capacity, count;
 } ValueArray;
 
 void value_array_init(ValueArray *array);
