+++ include/util/memory.h
@@ -8,7 +8,7 @@
 
 /**@desc get `object` byte located at `index`; `index` goes from LSB to MSB
 @return byte located at `index`*/
-inline uint8_t get_byte(unsigned long const object, int const index) {
+inline uint8_t get_byte(uint32_t const object, int const index) {
   assert(index >= 0 && "Expected index to be nonnegative");
   assert((size_t)index < sizeof(object) && "Expected index to fit within object (out of bounds)");
 
@@ -18,6 +18,6 @@ inline uint8_t get_byte(unsigned long const object, int const index) {
 typedef enum { LITTLE_ENDIAN, BIG_ENDIAN } Endianness;
 
 Endianness detect_endianness(void);
-unsigned long concatenate_bytes(int byte_count, ...);
+uint32_t concatenate_bytes(int byte_count, ...);
 
 #endif // MEMORY_H
