+++ src/util/memory.c
@@ -3,7 +3,7 @@
 
 #include "util/memory.h"
 
-uint8_t get_byte(unsigned long object, int index);
+uint8_t get_byte(uint32_t object, int index);
 
 /**@desc detect architecture endianness
 @return LITTLE_ENDIAN or BIG_ENDIAN*/
@@ -25,17 +25,18 @@ Endianness detect_endianness(void) {
 }
 
 /**@desc concatenate `byte_count` uint8_t `bytes`; `bytes` go from MSB to LSB
-@return unsigned long formed from `bytes` concatenation*/
-unsigned long concatenate_bytes(int byte_count, ...) {
+@return uint32_t formed from `bytes` concatenation*/
+uint32_t concatenate_bytes(int byte_count, ...) {
   assert(byte_count >= 2 && "Expected at least 2 bytes to concatenate");
   assert(
-    (size_t)byte_count <= sizeof(long) && "Exceeded number of bytes that can be concatenated without overflow"
+    (size_t)byte_count <= sizeof(uint32_t) &&
+    "Exceeded number of bytes that can be concatenated without overflow"
   );
 
   va_list bytes;
   va_start(bytes, byte_count);
 
-  unsigned long accumulator = 0;
+  uint32_t accumulator = 0;
   for (; byte_count > 0; byte_count--) {
     uint8_t const byte = va_arg(bytes, unsigned int);
     accumulator = (accumulator << CHAR_BIT) | byte;
